Organize email and SMS code a bit
