+++ packages/functions/src/index.ts
@@ -8,4 +8,6 @@ export * from "./testData";
 export * from "./testSlots";
 export * from "./https";
 export * from "./auth";
+
 export * from "./sendEmail";
+export * from "./sendSMS";
