+++ packages/functions/src/sendEmail/index.ts
@@ -1,0 +1,2 @@
+export * from "./https";
+export * from "./deliver";
